Comment out 'image' field in products table migration
The 'image' field in the products table migration has been commented out. This change was made as part of a larger refactor in the database schema. It prevents the 'image' field from being generated in the 'products' table for the time being as we reassess the need and implementation of this field.